+++ src/main/java/org/usf/inspect/core/HttpSessionStage.java
@@ -9,7 +9,7 @@ import java.util.UUID;
  */
 public final class HttpSessionStage extends AbstractStage {
 
-	public HttpSessionStage(UUID requestId, long order) {
+	public HttpSessionStage(UUID requestId, int order) {
 		super(requestId, order);
 	}
 }
